+++ packages/rake-db/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rake-db
 
+## 2.3.34
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.20
+
 ## 2.3.33
 
 ### Patch Changes